requirements.txt
beautifulsoup4==4.6.3
pycryptodome==3.7.2
